UI/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
Bottleneck==1.3.7
certifi==2024.2.2
click==8.1.7
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==2.2.5
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib_metadata==7.0.2
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest_asyncio==1.6.0
nltk==3.9.1
numexpr==2.8.7
numpy==1.26.4
openai==1.9.0
packaging==24.0
pandas==2.2.1
pandas-stubs==2.1.4.231227
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.42
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rouge-score==0.1.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.14.2
types-pytz==2022.4.0.0
typing_extensions==4.10.0
tzdata==2023.3
wcwidth==0.2.13
Werkzeug==2.3.8
wheel==0.41.2
zipp==3.17.0
gunicorn